initial abstract base class for installer and Cloud implementation
